Fix: change type of variable nthNumber to Long for overflow exception

+++ fibonacci/src/main/java/com/test/fibonacci/domain/model/Fibonacci.java
@@ -10,7 +10,7 @@ public class Fibonacci {
         this.fibonacciId=fibonacciId;
     }
 
-    public Integer getNthNumber() {
+    public Long getNthNumber() {
         return number.getNthNumber();
     }
